RAG_Basic_agentic_database/requirement.txt
streamlit
python-dotenv
requests
google-cloud-bigquery
db-dtypes
pandas
numpy